fix import modal not opening for other screens and auth checks in import touchpoints
